refactor/community_signin(profile 페이지(1))
