+++ demo250701/android/app/src/main/AndroidManifest.xml
@@ -3,6 +3,25 @@
         android:label="mvp"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
+        
+        <!-- Kakao SDK -->
+        <meta-data
+            android:name="com.kakao.vectormap.APP_KEY"
+            android:value="3c7a8b482a7de8109be0c367da2eb33a" />
+
+        <!-- Kakao OAuth Activity -->
+        <activity 
+            android:name="com.kakao.sdk.flutter.AuthCodeCustomTabsActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+                <data android:host="oauth"
+                      android:scheme="kakao3c7a8b482a7de8109be0c367da2eb33a" />
+            </intent-filter>
+        </activity>
+        
         <activity
             android:name=".MainActivity"
             android:exported="true"
