+++ demo250701/pubspec.yaml
@@ -46,7 +46,7 @@ dependencies:
   
   # Kakao Login SDK
   kakao_flutter_sdk: ^1.9.5
-  
+  kakao_flutter_sdk_user: ^1.9.5
   # HTTP requests for backend communication
   http: ^1.2.0
